Fire Detection
The main aim of this project is to detect fire and suppress it within time. These days the fire extinguisher system is in wide use for accident mitigation. The systems are so important that the occurrence of fatality can be reduced up to a greater extent, the following project is aiming to detect fire using a camera and the action of fire suppression will be taken on the location of fire.For the fire suppression the following system uses Water as an extinguishing material.The Automatic Fire Extinguisher System using a camera is an innovative and advanced solution designed to enhance fire detection and extinguishment in various environments. Traditional fire extinguishing systems often rely on manual intervention or simple sensors, which may lead to delayed response times or inaccuracies in detecting fires. The integration of a camera-based system brings a new level of efficiency and reliability to fire prevention